Fixes part of #10415: Topic validator refractor  (#10967)
* validators: Refactor topic and subtopic validators into individual validators

* validators: Refactor topic and resolve backend test coverage

* validators: Remove recommendations validator from topic validators file

* validators: Remove unused variable

* validators: Fix docstring naming and update copyright date

* validators: Define model vars in alphabetical order

* validators: Include topic test changes from oppia/oppia#10981

* Fix topic validators

* Fix backend tests

* Fix tests
